DRY SVG
